% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_replace}
\alias{find_replace}
\title{Find and Replace Text in an Rstudio Project Directory}
\usage{
find_replace(pattern, replacement, proj_dir = here::here(), ...)
}
\arguments{
\item{pattern}{Pattern to search for. Pattern is taken as a regular
expression (since it is passed to \code{gsub()}), so care needs to be taken when
passing this argument, both in terms of what you \emph{want} to change vs. what
you \emph{don't} want to change (e.g. \code{find_replace("ya.da", "yadda", "yabda")}
will return "yadda").}

\item{replacement}{Replacement for \code{pattern}.}

\item{proj_dir}{Project directory to search. By default, calls \code{\link[here:here]{here::here()}}
to guess.}

\item{...}{Other arguments passed on to \link[base:gsub]{gsub}.}
}
\description{
This function takes the current project directory (by default), and peforms a
project-wide, global find & replace in all files (via \link[base:gsub]{gsub()});
this is similar to how one might use
\href{https://www.gnu.org/software/sed/manual/sed.html}{sed}, but in R syntax.
This function is designed to be only run interactively, and is strict in its
request for confirmation, to prevent any accidental replacement.
}
\seealso{
\link[base:gsub]{gsub}
}
