% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%not.in\%}
\alias{\%not.in\%}
\title{Which elements of X are "not in" Y?}
\usage{
x \%not.in\% y
}
\arguments{
\item{x, y}{Vectors to compare. Order matters, just like when using
\link[base:match]{\%in\%}; for a symmetric version of
\link[base:setdiff]{setdiff()}, see \code{\link[=symdiff]{symdiff()}} in this package.}
}
\description{
Which elements of X are "not in" Y?
}
