% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is_equal}
\alias{is_equal}
\title{Vectorized Near-Equality}
\usage{
is_equal(x, y, tol = .Machine$double.eps)
}
\arguments{
\item{x, y}{Vectors of elements to compare.}

\item{tol}{Numeric tolerance for comparison. Default value should be
sufficient for most.}
}
\description{
This function serves as a vectorized alternative to
\link[base:all.equal]{all.equal()}.
}
