% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{symdiff}
\alias{symdiff}
\title{Symmetric Set Difference}
\usage{
symdiff(x, y)
}
\arguments{
\item{x, y}{Vectors to check for symmetric set difference.}
}
\description{
Like \link[base:setdiff]{setdiff()}, but symmetric! Not a new idea, but at
least exported here.
}
