% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{system_file}
\alias{system_file}
\title{Lightweight Locator for Package Files}
\usage{
system_file(file_spec)
}
\arguments{
\item{file_spec}{File specification}
}
\description{
Yadda. Add details.
}
\examples{
 \donttest{system_file("mypackage::shell/someShellScript.sh")}

}
